requirements.txt
bootstrap4==0.1.0
certifi==2022.5.18.1
chardet==3.0.4
coverage==6.4.1
distlib==0.3.4
dj-database-url==0.5.0
Django==2.0.4
django-adminlte2==0.4.1
django-filter==1.1.0
django-stdimage==5.0.1
djangorestframework==3.8.2
filelock==3.7.0
gunicorn==20.1.0
idna==2.8
importlib-metadata==4.11.4
model-mommy==2.0.0
mysqlclient==2.1.0
Pillow==9.1.1
platformdirs==2.5.2
progressbar2==4.0.0
psycopg2-binary==2.9.2
python-decouple==3.1
python-utils==3.2.3
pytz==2022.1
requests==2.21.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.24.3
virtualenv==20.14.1
zipp==3.8.0
